+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix NodePort with externaltrafficpolicy targets duplication @codearky
 - Update contributing section in README (#1760) @seanmalloy
 - Option to cache AWS zones list @bpineau
+- Refactor, enhance and test Akamai provider and documentation (#1846) @edglynes
 
 ## v0.7.3 - 2020-08-05